+++ src/de/omegazirkel/risingworld/tools/settings/AdminSettingsType.java
@@ -1,0 +1,7 @@
+package de.omegazirkel.risingworld.tools.settings;
+
+public enum AdminSettingsType {
+    BOOLEAN,
+    INTEGER,
+    STRING
+}